node: split openflowManager and service health check code
openflowManager code is only marginally health-check related, while the
service code provides healthz responses for local traffic policy
services on the node.

No code changes, just moving stuff around.
