+++ go-controller/pkg/node/healthcheck_service.go
@@ -3,22 +3,18 @@ package node
 import (
 	"fmt"
 	"net"
-	"os"
 	"sync"
-	"time"
 
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/config"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/factory"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/kube/healthcheck"
 	"github.com/ovn-org/ovn-kubernetes/go-controller/pkg/util"
-	"github.com/pkg/errors"
 
 	kapi "k8s.io/api/core/v1"
 	discovery "k8s.io/api/discovery/v1"
 	ktypes "k8s.io/apimachinery/pkg/types"
 	apierrors "k8s.io/apimachinery/pkg/util/errors"
 	"k8s.io/apimachinery/pkg/util/sets"
-	"k8s.io/klog/v2"
 	utilnet "k8s.io/utils/net"
 )
 
@@ -205,143 +201,6 @@ func isHostEndpoint(endpointIP string) bool {
 		}
 	}
 	return true
-}
-
-type openflowManager struct {
-	defaultBridge         *bridgeConfiguration
-	externalGatewayBridge *bridgeConfiguration
-	// flow cache, use map instead of array for readability when debugging
-	flowCache     map[string][]string
-	flowMutex     sync.Mutex
-	exGWFlowCache map[string][]string
-	exGWFlowMutex sync.Mutex
-	// channel to indicate we need to update flows immediately
-	flowChan chan struct{}
-}
-
-func (c *openflowManager) updateFlowCacheEntry(key string, flows []string) {
-	c.flowMutex.Lock()
-	defer c.flowMutex.Unlock()
-	c.flowCache[key] = flows
-}
-
-func (c *openflowManager) deleteFlowsByKey(key string) {
-	c.flowMutex.Lock()
-	defer c.flowMutex.Unlock()
-	delete(c.flowCache, key)
-}
-
-func (c *openflowManager) updateExBridgeFlowCacheEntry(key string, flows []string) {
-	c.exGWFlowMutex.Lock()
-	defer c.exGWFlowMutex.Unlock()
-	c.exGWFlowCache[key] = flows
-}
-
-func (c *openflowManager) requestFlowSync() {
-	select {
-	case c.flowChan <- struct{}{}:
-		klog.V(5).Infof("Gateway OpenFlow sync requested")
-	default:
-		klog.V(5).Infof("Gateway OpenFlow sync already requested")
-	}
-}
-
-func (c *openflowManager) syncFlows() {
-	// protect gwBridge config from being updated by gw.nodeIPManager
-	c.defaultBridge.Lock()
-	defer c.defaultBridge.Unlock()
-
-	c.flowMutex.Lock()
-	defer c.flowMutex.Unlock()
-
-	flows := []string{}
-	for _, entry := range c.flowCache {
-		flows = append(flows, entry...)
-	}
-
-	_, stderr, err := util.ReplaceOFFlows(c.defaultBridge.bridgeName, flows)
-	if err != nil {
-		klog.Errorf("Failed to add flows, error: %v, stderr, %s, flows: %s", err, stderr, c.flowCache)
-	}
-
-	if c.externalGatewayBridge != nil {
-		c.exGWFlowMutex.Lock()
-		defer c.exGWFlowMutex.Unlock()
-
-		flows := []string{}
-		for _, entry := range c.exGWFlowCache {
-			flows = append(flows, entry...)
-		}
-
-		_, stderr, err := util.ReplaceOFFlows(c.externalGatewayBridge.bridgeName, flows)
-		if err != nil {
-			klog.Errorf("Failed to add flows, error: %v, stderr, %s, flows: %s", err, stderr, c.exGWFlowCache)
-		}
-	}
-}
-
-// checkDefaultOpenFlow checks for the existence of default OpenFlow rules and
-// exits if the output is not as expected
-func (c *openflowManager) Run(stopChan <-chan struct{}, doneWg *sync.WaitGroup) {
-	doneWg.Add(1)
-	go func() {
-		defer doneWg.Done()
-		syncPeriod := 15 * time.Second
-		timer := time.NewTicker(syncPeriod)
-		defer timer.Stop()
-		for {
-			select {
-			case <-timer.C:
-				if err := checkPorts(c.defaultBridge.patchPort, c.defaultBridge.ofPortPatch,
-					c.defaultBridge.uplinkName, c.defaultBridge.ofPortPhys); err != nil {
-					klog.Errorf("Checkports failed %v", err)
-					continue
-				}
-				if c.externalGatewayBridge != nil {
-					if err := checkPorts(
-						c.externalGatewayBridge.patchPort, c.externalGatewayBridge.ofPortPatch,
-						c.externalGatewayBridge.uplinkName, c.externalGatewayBridge.ofPortPhys); err != nil {
-						klog.Errorf("Checkports failed %v", err)
-						continue
-					}
-				}
-				c.syncFlows()
-			case <-c.flowChan:
-				c.syncFlows()
-				timer.Reset(syncPeriod)
-			case <-stopChan:
-				return
-			}
-		}
-	}()
-}
-
-func checkPorts(patchIntf, ofPortPatch, physIntf, ofPortPhys string) error {
-	// it could be that the ovn-controller recreated the patch between the host OVS bridge and
-	// the integration bridge, as a result the ofport number changed for that patch interface
-	curOfportPatch, stderr, err := util.GetOVSOfPort("--if-exists", "get", "Interface", patchIntf, "ofport")
-	if err != nil {
-		return errors.Wrapf(err, "Failed to get ofport of %s, stderr: %q", patchIntf, stderr)
-
-	}
-	if ofPortPatch != curOfportPatch {
-		klog.Errorf("Fatal error: patch port %s ofport changed from %s to %s",
-			patchIntf, ofPortPatch, curOfportPatch)
-		os.Exit(1)
-	}
-
-	// it could be that someone removed the physical interface and added it back on the OVS host
-	// bridge, as a result the ofport number changed for that physical interface
-	curOfportPhys, stderr, err := util.GetOVSOfPort("--if-exists", "get", "interface", physIntf, "ofport")
-	if err != nil {
-		return errors.Wrapf(err, "Failed to get ofport of %s, stderr: %q", physIntf, stderr)
-	}
-	if ofPortPhys != curOfportPhys {
-		klog.Errorf("Fatal error: phys port %s ofport changed from %s to %s",
-			physIntf, ofPortPhys, curOfportPhys)
-		os.Exit(1)
-	}
-	return nil
 }
 
 func serviceNamespacedNameFromEndpointSlice(epSlice *discovery.EndpointSlice) (ktypes.NamespacedName, error) {
